refactor(ProjectTreeTable): extract DeleteConfirmContent into a separate component
- Moved `DeleteConfirmContent` to its own file
